requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.11.2
aiosignal==1.3.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.6.2.post1
asttokens==3.0.0
async-timeout==5.0.1
attrs==24.2.0
audioread==3.0.1
blinker==1.9.0
braceexpand==0.1.7
cachetools==5.5.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cloudpickle==3.1.0
contourpy==1.3.1
cycler==0.12.1
Cython==3.0.11
cytoolz==1.0.0
datasets==3.1.0
decorator==5.1.1
dill==0.3.8
Distance==0.1.3
docker-pycreds==0.4.0
docopt==0.6.2
editdistance==0.8.1
einops==0.8.0
exceptiongroup==1.2.2
executing==2.1.0
fastapi==0.115.5
fiddle==0.3.0
filelock==3.16.1
Flask==3.1.0
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2024.9.0
future==1.0.0
g2p-en==2.1.0
gitdb==4.0.11
GitPython==3.1.43
google-api-core==2.24.1
google-auth==2.38.0
google-cloud-speech==2.30.0
googleapis-common-protos==1.66.0
graphviz==0.20.3
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
huggingface_hub==0.26.2
hydra-core==1.3.2
idna==3.10
inflect==7.4.0
intervaltree==3.1.0
ipython==8.30.0
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.4
jiwer==3.0.5
joblib==1.4.2
kaldi-python-io==1.2.2
kaldiio==2.18.0
kiwisolver==1.4.7
lazy_loader==0.4
Levenshtein==0.26.1
lhotse==1.27.0
libcst==1.5.0
librosa==0.10.2.post1
lightning-utilities==0.11.8
lilcom==1.8.0
llvmlite==0.43.0
loguru==0.7.2
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.23.1
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.5.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.1.0
multiprocess==0.70.16
nemo_toolkit==2.0.0
networkx==3.4.2
nltk==3.9.1
numba==0.60.0
numpy==1.23.5
omegaconf==2.3.0
onnx==1.17.0
packaging==24.2
pandas==2.2.3
parso==0.8.4
pedalboard==0.9.16
pexpect==4.9.0
pillow==11.0.0
plac==1.4.3
platformdirs==4.3.6
pooch==1.8.2
prompt_toolkit==3.0.48
propcache==0.2.0
proto-plus==1.26.0
protobuf==5.28.3
psutil==6.1.0
ptyprocess==0.7.0
pure_eval==0.2.3
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyarrow==18.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
PyAudio==0.2.14
pybind11==2.13.6
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
Pygments==2.18.0
pyloudnorm==0.1.1
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
pyttsx3==2.98
pytz==2024.2
PyYAML==6.0.2
RapidFuzz==3.10.1
regex==2024.11.6
requests==2.32.3
resampy==0.4.3
rich==13.9.4
rsa==4.9
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
sacremoses==0.1.1
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.14.1
sentencepiece==0.2.0
sentry-sdk==2.18.0
setproctitle==1.3.3
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
soundfile==0.12.1
sox==1.5.0
soxr==0.5.0.post1
SpeechRecognition==3.11.0
stack-data==0.6.3
starlette==0.41.2
sympy==1.13.1
tabulate==0.9.0
tensorboard==2.18.0
tensorboard-data-server==0.7.2
termcolor==2.5.0
text-unidecode==1.3
texterrors==0.5.1
threadpoolctl==3.5.0
tokenizers==0.20.3
toolz==1.0.0
torch==2.5.1
torchaudio==2.5.1
torchmetrics==1.6.0
tqdm==4.67.0
traitlets==5.14.3
transformers==4.46.2
typeguard==4.4.1
typer==0.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
voicebox-tts==0.0.11
wandb==0.18.7
wcwidth==0.2.13
webdataset==0.2.100
Werkzeug==3.1.3
wget==3.2
wrapt==1.16.0
xxhash==3.5.0
yarl==1.17.1
